task 2 done, project structure organized
