+++ src/main/java/com/agency04/sbss/pizza/model/pizzaPojos/QuatroFormaggi.java
@@ -1,24 +1,28 @@
-package com.agency04.sbss.pizza.pizzaPojos;
+package com.agency04.sbss.pizza.model.pizzaPojos;
 
-import com.agency04.sbss.pizza.Pizza;
-import com.agency04.sbss.pizza.PizzaIngredient;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Calzone implements Pizza {
+@Component
+public class QuatroFormaggi  implements Pizza {
     private List<PizzaIngredient> ingredients = new ArrayList<>();
 
-    public Calzone(){
+    @Value("${quatro.name}")
+    private String name;
+
+    public QuatroFormaggi(){
         ingredients.add(PizzaIngredient.TOMATO_SAUCE);
         ingredients.add(PizzaIngredient.MOZZARELLA);
-        ingredients.add(PizzaIngredient.HAM);
-        ingredients.add(PizzaIngredient.ANCHOVIES);
+        ingredients.add(PizzaIngredient.PARMESAN);
+        ingredients.add(PizzaIngredient.GORGONZOLA);
         ingredients.add(PizzaIngredient.OREGANO);
     }
     @Override
     public String getName() {
-        return "Calzone";
+        return this.name;
     }
 
     @Override
