+++ src/main/java/com/agency04/sbss/pizza/model/pizzeriaPojos/PizzeriaService.java
@@ -1,4 +1,6 @@
-package com.agency04.sbss.pizza;
+package com.agency04.sbss.pizza.model.pizzeriaPojos;
+
+import com.agency04.sbss.pizza.model.pizzaPojos.Pizza;
 
 public interface PizzeriaService {
     public String getName();
