+++ src/main/java/com/agency04/sbss/pizza/service/PizzaDeliveryService.java
@@ -1,0 +1,25 @@
+package com.agency04.sbss.pizza.service;
+
+import com.agency04.sbss.pizza.model.pizzaPojos.Pizza;
+import com.agency04.sbss.pizza.model.pizzeriaPojos.PizzeriaService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Component;
+
+@Component
+public class PizzaDeliveryService {
+
+    @Qualifier("dominos")
+    @Autowired
+    private PizzeriaService pizzeriaService;
+
+    @Autowired
+    public PizzaDeliveryService(){
+    }
+
+    public String orderPizza(Pizza pizza) throws InterruptedException {
+        pizzeriaService.makePizza(pizza);
+        Thread.sleep(2000);
+        return pizza.getName() +" is on it's way :) !";
+    }
+}
