+++ src/main/resources/application.properties
@@ -1,2 +1,10 @@
-foo.name=Dominos
-foo.address= Andrije Žaje 60
+dominos.name = Dominos
+dominos.address = Andrije Žaje 60
+
+lastrada.name = Pizzeria Lastrada
+lastrada.address = Obala kneza Domagoja 56
+
+vegeteriana.name = Vegeteriana
+margherita.name = Margherita
+quatro.name = Quatro Formaggi
+calzone.name = Calzone
