+++ src/main/resources/spring-context.xml
@@ -7,26 +7,10 @@
        http://www.springframework.org/schema/context
        http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <!-- load the properties file -->
-    <context:property-placeholder location="classpath:application.properties"/>
-    <!-- Define your Beans here -->
-    <bean id="pizzeriaLastrada"
-          class="com.agency04.sbss.pizza.pizzeriaPojos.PizzeriaLastrada">
-        <property name="name" value="Pizzeria Lastrada"/>
-        <property name="address" value="Obala kneza Domagoja 56"/>
-    </bean>
+       <!-- load the properties file -->
+        <context:property-placeholder location="classpath:application.properties"/>
 
-    <bean id="Dominos"
-          class="com.agency04.sbss.pizza.pizzeriaPojos.Dominos">
-          <property name="name" value="${foo.name}"/>
-          <property name="address" value="${foo.address}"/>
-    </bean>
-
-    <bean id="pizzaDeliveryService"
-          class="com.agency04.sbss.pizza.PizzaDeliveryService">
-        <!--> <constructor-arg ref="pizzeriaLastrada"/>  <-->
-        <property name="pizzeriaService" ref="Dominos"/>
-    </bean>
-
+       <!-- add entry to enable component scanning -->
+        <context:component-scan base-package="com.agency04.sbss.pizza"/>
 
 </beans>
